\docType{data}
\name{calibrationSample}
\alias{calibrationSample}
\title{Calibration sample data}
\description{
  #' This includes the calibration-sample data for the
  international conflict forecasting example in Montgomery,
  Hollenbach and Ward (2012). It provides the predictions
  for the three models included in the Ensemble models, as
  well as the true values of the dependent variable for
  insurgency in 29 Asian countries. The calibration sample
  ranges ranges from January 2008 to December 2009.
}
\keyword{datasets}

