\docType{package}
\name{EBMAforecast}
\alias{EBMAforecast}
\alias{EBMAforecast-package}
\title{EBMAforecast}
\description{
  The EBMAforecast package (currently under development)
  allows users to increase the accuracy of forecasting
  models by pooling multiple component forecasts to
  generate ensemble forecasts. It includes functions to fit
  an ensemble Bayesian model averaging (EBMA) model using
  in-sample predictions, generate ensemble out-of-sample
  predictions, and create useful data visualizations
}
\examples{
demo(EBMAforecast)
}
\author{
  Michael D. Ward and Jacob M. Montgomery
}
\references{
  Montgomery, Jacob M., Florian M. Hollenbach and Michael
  D. Ward. (2012). Improving Predictions Using Ensemble
  Bayesian Model Averaging. \emph{Political Analysis}.
  Forthcoming.
}

