\name{plot.Ensemble.logit}
\alias{plot.Ensemble.logit}
\title{Plotting function for logistic ensemble models that have already been fit}
\usage{
  \method{print}{FDatFitLogit} (obj)
}
\arguments{
  \item{.object}{An object of class "FDatFitLogit"}

  \item{.period}{"calibration" or "test"}
}
\description{
  Description goes here
}

