\docType{data}
\name{testSample}
\alias{testSample}
\title{Test sample data}
\description{
  This includes the test-sample data for the international
  conflict forecasting example in Montgomery, Hollenbach
  and Ward (2012). It provides the predictions for the
  three models included in the Ensemble models, as well as
  the true values of the dependent variable for insurgency
  in 29 Asian countries. The test sample ranges ranges from
  January 2010 to December 2010.
}
\keyword{datasets}

