% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EBMAforecast-package.R
\docType{package}
\name{EBMAforecast-package}
\alias{EBMAforecast}
\alias{EBMAforecast-package}
\title{EBMAforecast: Estimate Ensemble Bayesian Model Averaging Forecasts using Gibbs Sampling or EM-Algorithms}
\description{
Create forecasts from multiple predictions using ensemble Bayesian model averaging (EBMA). EBMA models can be estimated using an expectation maximization (EM) algorithm or as fully Bayesian models via Gibbs sampling. The methods in this package are Montgomery, Hollenbach, and Ward (2015) \doi{10.1016/j.ijforecast.2014.08.001} and Montgomery, Hollenbach, and Ward (2012) \doi{10.1093/pan/mps002}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fhollenbach/EBMA/}
}

}
\author{
\strong{Maintainer}: Florian M. Hollenbach \email{fho.egb@cbs.dk} (\href{https://orcid.org/0000-0002-9599-556X}{ORCID})

Authors:
\itemize{
  \item Jacob M. Montgomery
  \item Michael D. Ward
}

}
\keyword{internal}
